+++ structures/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "structures.apps.StructuresConfig"
 
-__version__ = "1.20.5"
+__version__ = "1.21.0a1"
 __title__ = "Structures"