+++ structures/models/owners.py
@@ -1047,7 +1047,7 @@ class Owner(models.Model):
             # at least one type has a trailing white space
             # which we need to remove
             notif_type = notification["type"].strip()
-            notif = Notification.objects.create(
+            Notification.objects.create(
                 notification_id=notification["notification_id"],
                 owner=self,
                 sender=sender,
@@ -1058,8 +1058,6 @@ class Owner(models.Model):
                 last_updated=now(),
                 created=now(),
             )
-            if notif.is_structure_related:
-                notif.update_related_structures()
         return len(new_notifications)
 
     def _process_timers_for_notifications(self, token: Token):